moved navdrawer stuff to mainactivity, still not working as intended
