+++ app/src/main/AndroidManifest.xml
@@ -14,20 +14,13 @@
         <activity
             android:name=".MainActivity"
             android:screenOrientation="portrait">
-        </activity>
-        <activity android:name=".LoginActivity" />
-        <activity
-            android:name=".NavigationDrawerActivity"
-            android:label="@string/title_activity_navigation_drawer"
-            android:parentActivityName=".MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
-            <meta-data
-                android:name="android.support.PARENT_ACTIVITY"
-                android:value="com.greenfox.tribesoflagopusandroid.MainActivity" />
         </activity>
+        <activity android:name=".LoginActivity" />
+
     </application>
 
 </manifest>
