+++ app/src/main/java/com/greenfox/tribesoflagopusandroid/MainActivity.java
@@ -5,7 +5,15 @@ import android.content.Intent;
 import android.content.SharedPreferences;
 import android.os.Bundle;
 import android.preference.PreferenceManager;
+import android.support.annotation.NonNull;
+import android.support.design.widget.NavigationView;
+import android.support.v4.app.Fragment;
+import android.support.v4.app.FragmentTransaction;
+import android.support.v4.view.GravityCompat;
+import android.support.v4.widget.DrawerLayout;
+import android.support.v7.app.ActionBarDrawerToggle;
 import android.support.v7.app.AppCompatActivity;
+import android.support.v7.widget.Toolbar;
 import android.view.Menu;
 import android.view.MenuInflater;
 import android.view.MenuItem;
@@ -14,7 +22,10 @@ import android.view.View;
 import android.widget.Button;
 import android.widget.Toast;
 
-public class MainActivity extends AppCompatActivity {
+import com.greenfox.tribesoflagopusandroid.fragments.BattleFragment;
+import com.greenfox.tribesoflagopusandroid.fragments.BuildingsFragment;
+
+public class MainActivity extends AppCompatActivity implements NavigationView.OnNavigationItemSelectedListener {
 
     public static final String USERNAME = "Username";
     public static final String PASSWORD = "Password";
@@ -22,11 +33,25 @@ public class MainActivity extends AppCompatActivity {
     SharedPreferences preferences;
     SharedPreferences.Editor editor;
 
+    private DrawerLayout mDrawer;
+
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_main);
         checkUsernameAndPassword();
+
+        setContentView(R.layout.activity_navigation_drawer);
+        Toolbar toolbar = (Toolbar) findViewById(R.id.toolbar);
+        setSupportActionBar(toolbar);
+
+        DrawerLayout drawer = (DrawerLayout) findViewById(R.id.drawer_layout);
+        ActionBarDrawerToggle toggle = new ActionBarDrawerToggle(
+                this, drawer, toolbar, R.string.navigation_drawer_open, R.string.navigation_drawer_close);
+        drawer.setDrawerListener(toggle);
+        toggle.syncState();
+        NavigationView navigationView = (NavigationView) findViewById(R.id.nav_view);
+        navigationView.setNavigationItemSelectedListener(this);
 
         Button button = (Button) findViewById(R.id.logout);
         button.setOnClickListener(new View.OnClickListener() {
@@ -42,12 +67,16 @@ public class MainActivity extends AppCompatActivity {
         getMenuInflater().inflate(R.menu.navigation_drawer, menu);
         return true;
     }
+
     @Override
     public boolean onOptionsItemSelected(MenuItem item) {
         super.onOptionsItemSelected(item);
-        switch(item.getItemId()) {
+        switch (item.getItemId()) {
             case R.id.refreshing:
                 Toast.makeText(MainActivity.this, "Refreshing", Toast.LENGTH_SHORT).show();
+            case R.id.home:
+                mDrawer.openDrawer(GravityCompat.START);
+                return true;
         }
         return false;
     }
@@ -72,4 +101,37 @@ public class MainActivity extends AppCompatActivity {
         finish();
     }
 
+    private void displaySelectedScreen(int id) {
+        Fragment fragment = null;
+
+        switch (id) {
+            case R.id.nav_buildings:
+                fragment = new BuildingsFragment();
+                break;
+            case R.id.nav_kingdom:
+                break;
+            case R.id.nav_battle:
+                fragment = new BattleFragment();
+                break;
+            case R.id.nav_settings:
+                break;
+            case R.id.nav_troops:
+                break;
+        }
+        if (fragment != null) {
+            FragmentTransaction transaction = getSupportFragmentManager().beginTransaction();
+            transaction.replace(R.id.imageView, fragment);
+            transaction.commit();
+        }
+        DrawerLayout drawer = (DrawerLayout) findViewById(R.id.drawer_layout);
+        drawer.closeDrawer(GravityCompat.START);
+    }
+
+    @SuppressWarnings("StatementWithEmptyBody")
+    @Override
+    public boolean onNavigationItemSelected(MenuItem item) {
+        int id = item.getItemId();
+        displaySelectedScreen(id);
+        return true;
+    }
 }
